refactor: rename files for consistency
