+++ src/index.ts
@@ -15,4 +15,4 @@ export { RouterLink, RouterLinkProps } from './components/router-link';
 export { HistoryAdapter } from './adapters/history-adapter';
 export { StaticAdapter } from './adapters/static-adapter';
 export { generateUrl, routerStateToUrl } from './adapters/generate-url';
-export { findMatchingRoute } from './adapters/match-current-route';
+export { findMatchingRoute } from './adapters/find-matching-route';
